Refactor client side RPC
that the generic RPC functionality is contained within client_rpc.ts
and called from the other scripts. Added dumb_server.ts as an example
of a super simple rpc implementation to a simple server. Cleaned out
registry.ts so that only specific code relating the the implementation
of that RPC is present.