Change jsonrpc macros to take textual route defs instead of annotations.
This lets us remove the Call and Notify annotations which were causing
problems depending on the naming of files and consequent order of
compilation, which is just weird. By removing the annotation classes,
we make the api classes completely independent of jsonrpc, which is a
nice feature. All the knowledge of jsonrpc is contained in the ApiBackend
class where we define routes for both the client to server and server to
client directions. This is very simlar to the system play uses for defining
a separate routes file except that we use macros instead of code generation.

Review: @joshmutus