Use Activity for managing url transitions and add basic grapher plot UIs
The Activity class is where we perform asynchronous initialization and cleanup
for screens in the app, since they may require server communication before the ui
can be rendered. This will allow us to keep logic out of the polymer element
classes as much as possible, so they can focus on visual representation of data.

Plotting is done using d3 to create svg documents. We support simple 1D and 2D
plots without a lot of visual customization yet, but that can come later.

Review: @joshmutus