Clean up registry copy logic to not interleave reading and writing.
We previously changed the dir copying code to avoid a potential
infinite recursion due to copying a folder directly inside itself,
however the fix did not handle a case when a folder is copied into
a subfolder of itself. This fixes these sorts of problems by first
loading the entire directory contents, including all subdirectory
contents, and then writing the whole thing to the new location
in one shot. This is more memory intensive on the server, but for
typical copies that we do it should be fine.

Review: @DanielSank (OTS), @joshmutus (testing)
Fixes #124