Use base tag to allow relative urls in index.html
We send index.html to the browser to bootstrap the app no matter what
route within the app the user loads. Hence, to use relative urls we must
update the base against which such urls are resolved based on the root
where the app is hosted, not the route from which the page was served.
On the server, we modify the base href appropriately before serving the
page by manipulating the page DOM (previously this was done by injecting
javascript, but the DOM manipulation is much cleaner).