Allow specifying manager suffix that will apply to all manager hosts.
We currently allow specifying aliases for manager hosts that the web
interface can connect to so that one can navigate to /foo/ and connect
to a manager running on foo.example.com, say. This change allows
specifying a single hostname suffix, rather than requiring all such
aliases to be specified explicitly.