Created an IronIconBakery and ListBuilder class to allow for more performant UI list elements. Applied these classes to the Grapher directory/dataset listing.
Polymer does a lot of fancy databinding and event listening, which makes things
very slow at scale. Rendering a list of 4000 datasets would freeze the entire
application for 6-7 seconds, sometimes more. Even traversing simple lists took
a noticable amount of time which just leads to a frustrating user experience,
especially if navigating multiple levels of a directory structure.

To fix this, I've moved the rendering of the grapher list view into raw
JavaScript rather than using the Polymer templating language. As a result,
small lists now render instantly, and large lists, while still take a few
seconds to entirely load (the DOM just doesn't like this many elements being
added) the page and the top of the list renders instantly giving the user the
impression of speed.

This PR is a first step towards overhauling and unifying the list views across
the application to allow for easier fixes and improvements to the UI.