Use iron-list for displaying menu in grapher. (#206)
Updated the menu listing in the Grapher UI to use the `iron-list` component provided by Polymer. This greatly improves scrolling performance on large lists, as well as speeds up page loading times considerably.

This required changing the the grapher activity to pass only actual updates to the view rather than re-syncing the entire list via the API constantly. This is the preferred design pattern moving forwards.

It also moves the UI design away from `table` and towards the use of `div`, which was a happy byproduct of how `iron-list` operates.