Add a color bar axis to the Grapher for 2D plots. (#219)
Fixes issue #90 where there was no axis for the colors displayed on 2D plots using the Grapher.

Added a new element to 2D plots showing the z-axis as a color gradient. The scale adjusts as new data is introduced. Size and positioning can be tweaked via constants in the file.

Also moved the styling for the zoomRectangle from global CSS, to a local fill color on the SVG element. This was done because otherwise all rectangles had to be the same color.