WebGL Implementation of main plotting interface in Grapher. (#223)
Resolves issues #185, #208, #231. Makes use of WebGL to do the data plotting portion of each graph in order to greatly improve the performance of live data view as well as zoom/pan operations.

d3.js was retained for the axes and and interface as SVG, while THREE.js library was introduced to handle the plotting of the data via WebGL.

This dual library usage was used to get the most performance gain for the least development time, as implementing axes and other features in WebGL will be time consuming. It is a longer term goal to move all plotting towards WebGL in order to allow screenshotting (Issue #184).

Performance was also improved by causing only the parts of the graph that need to be re-rendered under actions like zooming to be re-rendered. This also makes it possible to now retain the zoom level during live updates, the changing of plotting types, and the toggling of traces.