Fix registry modal sizing when working with very long keys. (#246)
Fixes #244 where very long registry keys would scroll off the page due to the dialog not resizing properly.

This PR also fixes the unreported issue that if you pasted a long key into the new key or edit key windows, the modal window would not reposition, causing it to go off screen.

This bug was caused by an interplay of the various Polymer elements being used. The main culprit being that the `iron-resize` event that `paper-dialog` listens to so that it knows when to resize/reposition does not get fired by a `iron-autogrow-textarea` which is a lower level component that `paper-textarea` wraps.

This is resolved by first simplifying the component interaction stack by removing the use on `paper-dialog-scrollable` and moving the job of scrolling to the dialog itself. Secondly, the value update event on the textarea is hooked, and used to fire an `iron-resize` event that the dialog can catch, causing both the position and sizing to be set correctly.

Review: @maffoo