Show exceptions on Nodes when they occur during start, stop or restart of a server. (#250)
Fixes #169. Shows the user that an error has occurred if an exception is thrown during the starting, stopping or resetting of a server from the Node panel. Also show the whole exception that was raised in a collapsible panel.

Previously, when an exception occurred while starting, stopping or resetting a server, it was logged to the console, but not displayed to the user.

Also adjusts the Node UI to fill the screen and fixes the "jump" that would occur when the loading spinner was active on a line of controls.

Review: @maffoo