[#143, Part 2] Added filtering of servers in Node view based on autostart status. (#257)
Partial fix for #143. Added controls for filtering the list of servers based on if they appear in an autostart list or not. 

This required refactoring the code to no longer use computed properties. Specifically, with the original way, when filtering the list, all state information of the servers would be lost as the list would just get recreated from scratch. The new code isn't substantially more complicated than the existing, despite having to properly keep the list state current.

This also fixes an unreported bug where the control states were incorrect on page load due to the server controls not communicating their state to one another.

Review: @maffoo