[#182] Add functionality to fit Parabolic and Exponential functions to plots. (#262)
Implements #182. Added functionality to fit Parabolic and Exponential functions to plots. 

Added a new 'Fit' mode that allows the drawing of a bounding box. Data within this bounding box will be used to fit either a Parabola or Exponential (dependent on the value of the drop down menu added to the interface). Coefficients of the fits, as well as T1 for Exponential, and the extremum for Parabolic are shown.

Review: @maffoo