[#278] Change Registry to use iron-list. (#280)
Implements #278. Updated/Refactored the Registry to use an `iron-list` instead of a `table` with `dom-repeat`. This allows for smooth scrolling on the registry when using keyboard controls.

This also refactored the code to get rid of all computed properties, as well as the last instance of the `kick` hack. Also fixed problems with submitting forms via hitting enter in text fields, or while OK is selected. In addition, various style guide compliance changes.

Review: @maffoo