Added a throttle to selectedDataset updates on Grapher. (#292)
Currently every time you select an item in the Grapher, the `selectedDataset` computed property makes a call to the server and then renders all the info about that dataset. With the introduction of keyboard controls, holding a key leads to a very rapid shuffle through the selected datasets which hammers the server and lags the UI due to all the DOM updates.

This PR fixes that by adding a 100ms throttle to the `selectedDataset` update. The call to the server is delayed 100ms, and any future calls made within that window will cancel the request and create a new one. This means that you only call the server at the end of a scroll, making the UI much snappier and reducing calls to the server drastically.

Review: @maffoo