Fix splicing of lastData row in grapher
The second argument to splice is the number of items to delete from the
array. When splicing in the lastData row we were deleting the next row,
and if there was only one row to begin with this would result in a data
array of length 1, causing createLine to return null and the data point
not to be plotted. This problem becomes more pronounced when the data
arrives slowly, one row at a time, as often happens when taking real-
world experimental data. Added a slow demo scan to test this scenario.

Fixes #298