Get bounding rect each time we recompute mouse position.
Previously this was computed once when the plot was attached to the
document and subsequently any time it was resized. However, it also has
to be recomputed whenever the plot element changes position in the
document (e.g. in live view when other plots are added) and also
whenever the document is scrolled (since the bounding rect is relative
to the viewport and this changes under scrolling). I don't notice any
particular performance problem due to this recomputation even with
vigorous, continuous mouse movement over a plot; I think simplicity
and correctness are worth the small performance hit.

Fixes #303