Fix a few bugs in plot element (#307)
* Do not add null lines to plot.
* Deal with 0-dimensional data sets.
* Get bounding rect each time we recompute mouse position.

Without the check we get errors later when trying to add the lines to the three.js scene object.

Previously, the bounding rect was computed once when the plot was attached to the document and subsequently any time it was resized. However, it also has to be recomputed whenever the plot element changes position in the document (e.g. in live view when other plots are added) and also whenever the document is scrolled (since the bounding rect is relative to the viewport and this changes under scrolling). Performance does not seem to suffer due to this recomputation even with vigorous, continuous mouse movement over a plot; simplicity and correctness are certainly improved.

Fixes #303, Fixes #306 
Review: @kunalq